andrew-ar/requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
